addressed comments in PR: move file, update comment, filter out more options, remove redundant warning
