+++ src/validation/__tests__/KnownArgumentNames-test.js
@@ -109,7 +109,7 @@ describe('Validate: Known argument names', () => {
         dog @skip(unless: true)
       }
     `, [
-      unknownDirectiveArg('unless', 'skip', [ 'if' ], 3, 19),
+      unknownDirectiveArg('unless', 'skip', [], 3, 19),
     ]);
   });
 
@@ -119,7 +119,7 @@ describe('Validate: Known argument names', () => {
         doesKnowCommand(unknown: true)
       }
     `, [
-      unknownArg('unknown', 'doesKnowCommand', 'Dog', [ 'dogCommand' ], 3, 25),
+      unknownArg('unknown', 'doesKnowCommand', 'Dog', [], 3, 25),
     ]);
   });
 
@@ -129,8 +129,8 @@ describe('Validate: Known argument names', () => {
         doesKnowCommand(whoknows: 1, dogCommand: SIT, unknown: true)
       }
     `, [
-      unknownArg('whoknows', 'doesKnowCommand', 'Dog', [ 'dogCommand' ], 3, 25),
-      unknownArg('unknown', 'doesKnowCommand', 'Dog', [ 'dogCommand' ], 3, 55),
+      unknownArg('whoknows', 'doesKnowCommand', 'Dog', [], 3, 25),
+      unknownArg('unknown', 'doesKnowCommand', 'Dog', [], 3, 55),
     ]);
   });
 
@@ -149,8 +149,8 @@ describe('Validate: Known argument names', () => {
         }
       }
     `, [
-      unknownArg('unknown', 'doesKnowCommand', 'Dog', [ 'dogCommand' ], 4, 27),
-      unknownArg('unknown', 'doesKnowCommand', 'Dog', [ 'dogCommand' ], 9, 31),
+      unknownArg('unknown', 'doesKnowCommand', 'Dog', [], 4, 27),
+      unknownArg('unknown', 'doesKnowCommand', 'Dog', [], 9, 31),
     ]);
   });
